Creating new DEV environment variable
