+++ src/interfaces/response/index.ts
@@ -1,0 +1,5 @@
+export interface IResponse<T> {
+    content: T;
+    httpStatus: number;
+    errorMessage: string;
+}
