+++ src/services/index.ts
@@ -1,0 +1,5 @@
+import { IResponse } from '../interfaces/response';
+
+export interface ILoginService {
+    login(name: string, password: string): Promise<IResponse<boolean>>;
+}
